+++ monitor/price/pricemonitoring_test.go
@@ -48,7 +48,7 @@ func TestEmptyParametersList(t *testing.T) {
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -70,6 +70,8 @@ func TestEmptyParametersList(t *testing.T) {
 }
 
 func TestErrorWithNilRiskModel(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
 	t1 := proto.PriceMonitoringTrigger{Horizon: 7200, Probability: "0.95", AuctionExtension: 300}
 	t2 := proto.PriceMonitoringTrigger{Horizon: 3600, Probability: "0.99", AuctionExtension: 60}
 
@@ -78,11 +80,11 @@ func TestErrorWithNilRiskModel(t *testing.T) {
 			Triggers: []*proto.PriceMonitoringTrigger{&t1, &t2},
 		},
 	}
+	auctionStateMock := mocks.NewMockAuctionState(ctrl)
 	settings := types.PriceMonitoringSettingsFromProto(pSet)
-	ctrl := gomock.NewController(t)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	// statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
-	pm, err := price.NewMonitor("asset", "market", nil, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", nil, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.Error(t, err)
 	require.Nil(t, pm)
 }
@@ -90,6 +92,7 @@ func TestErrorWithNilRiskModel(t *testing.T) {
 func TestGetHorizonYearFractions(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	riskModel := mocks.NewMockRangeProvider(ctrl)
+	auctionStateMock := mocks.NewMockAuctionState(ctrl)
 	t1 := proto.PriceMonitoringTrigger{Horizon: 7200, Probability: "0.95", AuctionExtension: 300}
 	t2 := proto.PriceMonitoringTrigger{Horizon: 3600, Probability: "0.99", AuctionExtension: 60}
 
@@ -101,7 +104,7 @@ func TestGetHorizonYearFractions(t *testing.T) {
 	settings := types.PriceMonitoringSettingsFromProto(pSet)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -131,10 +134,12 @@ func TestRecordPriceChange(t *testing.T) {
 
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(4)
 	auctionStateMock.EXPECT().InAuction().Return(false).Times(4)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(2)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -187,10 +192,12 @@ func TestCheckBoundViolationsWithinCurrentTimeWith2HorizonProbabilityPairs(t *te
 
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(16)
 	auctionStateMock.EXPECT().InAuction().Return(false).Times(16)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).Times(12)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(12)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -231,7 +238,7 @@ func TestCheckBoundViolationsWithinCurrentTimeWith2HorizonProbabilityPairs(t *te
 	require.False(t, b)
 
 	// Reinstantiate price monitoring after auction to reset internal state
-	pm, err = price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err = price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -248,7 +255,7 @@ func TestCheckBoundViolationsWithinCurrentTimeWith2HorizonProbabilityPairs(t *te
 	require.False(t, b)
 
 	// Reinstantiate price monitoring after auction to reset internal state
-	pm, err = price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err = price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -264,7 +271,7 @@ func TestCheckBoundViolationsWithinCurrentTimeWith2HorizonProbabilityPairs(t *te
 	require.False(t, b)
 
 	// Reinstantiate price monitoring after auction to reset internal state
-	pm, err = price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err = price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -280,7 +287,7 @@ func TestCheckBoundViolationsWithinCurrentTimeWith2HorizonProbabilityPairs(t *te
 	require.False(t, b)
 
 	// Reinstantiate price monitoring after auction to reset internal state
-	pm, err = price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err = price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -296,7 +303,7 @@ func TestCheckBoundViolationsWithinCurrentTimeWith2HorizonProbabilityPairs(t *te
 	require.False(t, b)
 
 	// Reinstantiate price monitoring after auction to reset internal state
-	pm, err = price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err = price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -559,11 +566,12 @@ func TestAuctionStartedAndEndendBy1Trigger(t *testing.T) {
 
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(2)
 	auctionStateMock.EXPECT().InAuction().Return(false).Times(2)
-	auctionStateMock.EXPECT().IsPriceAuction().Return(true).AnyTimes()
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).Times(2)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	pm.UpdateTestFactors(downFactorsP1, upFactorsP1)
@@ -586,6 +594,8 @@ func TestAuctionStartedAndEndendBy1Trigger(t *testing.T) {
 
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(1)
 	auctionStateMock.EXPECT().InAuction().Return(true).Times(1)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(true).Times(1)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(true).Times(1)
 	auctionStateMock.EXPECT().IsOpeningAuction().Return(false).Times(1)
 	// auctionStateMock.EXPECT().IsPriceAuction().Return(true).Times(1)
 	auctionStateMock.EXPECT().ExpiresAt().Return(&initialAuctionEnd).Times(1)
@@ -619,10 +629,12 @@ func TestAuctionStartedAndEndendBy2Triggers(t *testing.T) {
 
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(2)
 	auctionStateMock.EXPECT().InAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(4)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).Times(3)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -648,6 +660,7 @@ func TestAuctionStartedAndEndendBy2Triggers(t *testing.T) {
 	auctionStateMock.EXPECT().InAuction().Return(true).Times(1)
 	auctionStateMock.EXPECT().IsOpeningAuction().Return(false).Times(1)
 	auctionStateMock.EXPECT().IsPriceAuction().Return(true).Times(1)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).Times(1)
 	auctionStateMock.EXPECT().ExpiresAt().Return(&initialAuctionEnd).Times(1)
 	auctionStateMock.EXPECT().SetReadyToLeave().Times(1)
 
@@ -689,10 +702,12 @@ func TestAuctionStartedAndEndendBy1TriggerAndExtendedBy2nd(t *testing.T) {
 
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(2)
 	auctionStateMock.EXPECT().InAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).AnyTimes()
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	downFactors := []num.Decimal{pMin1.Div(decPrice), pMin2.Div(decPrice)}
 	upFactors := []num.Decimal{pMax1.Div(decPrice), pMax2.Div(decPrice)}
 	pm.UpdateTestFactors(downFactors, upFactors)
@@ -780,6 +795,103 @@ func TestAuctionStartedAndEndendBy1TriggerAndExtendedBy2nd(t *testing.T) {
 	require.False(t, b)
 }
 
+func TestAuctionStartedBy1TriggerAndNotExtendedBy2ndStaleTrigger(t *testing.T) {
+	// Also verifies that GetCurrentBounds() works as expected
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	riskModel := mocks.NewMockRangeProvider(ctrl)
+	auctionStateMock := mocks.NewMockAuctionState(ctrl)
+	price1 := num.NewUint(123)
+	cp1 := []*types.Trade{{Price: price1, Size: 1}}
+	now := time.Date(1993, 2, 2, 6, 0, 0, 1, time.UTC)
+	t1 := proto.PriceMonitoringTrigger{Horizon: 6, Probability: "0.95", AuctionExtension: 60}
+	t2 := proto.PriceMonitoringTrigger{Horizon: 6, Probability: "0.99", AuctionExtension: 120}
+	pSet := &proto.PriceMonitoringSettings{
+		Parameters: &proto.PriceMonitoringParameters{
+			Triggers: []*proto.PriceMonitoringTrigger{&t1, &t2},
+		},
+	}
+	settings := types.PriceMonitoringSettingsFromProto(pSet)
+	ctx := context.Background()
+	decPrice, pMin1, pMax1, _, maxUp1 := getPriceBounds(price1, 1, 2)
+	_, pMin2, pMax2, _, maxUp2 := getPriceBounds(price1, 1*4, 2*4)
+
+	one := num.NewUint(1)
+	t1lb1, _ := num.UintFromDecimal(pMin1)
+	t1lb1.AddSum(one) // account for value being ceil'ed
+	t1ub1, _ := num.UintFromDecimal(pMax1)
+	t1ub1.Sub(t1ub1, one) // floor
+	t2lb1, _ := num.UintFromDecimal(pMin2)
+	t2lb1.AddSum(one) // again: ceil
+	t2ub1, _ := num.UintFromDecimal(pMax2)
+
+	auctionStateMock.EXPECT().IsFBA().Return(false).Times(2)
+	auctionStateMock.EXPECT().InAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(2)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).AnyTimes()
+	statevar := mocks.NewMockStateVarEngine(ctrl)
+	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
+
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
+	downFactors := []num.Decimal{pMin1.Div(decPrice), pMin2.Div(decPrice)}
+	upFactors := []num.Decimal{pMax1.Div(decPrice), pMax2.Div(decPrice)}
+	pm.UpdateTestFactors(downFactors, upFactors)
+
+	require.NoError(t, err)
+	require.NotNil(t, pm)
+
+	pm.OnTimeUpdate(now)
+	b := pm.CheckPrice(ctx, auctionStateMock, cp1, true)
+	require.False(t, b)
+
+	bounds := pm.GetCurrentBounds()
+	require.Len(t, bounds, 2)
+	require.Equal(t, *bounds[0].Trigger.IntoProto(), t1)
+	require.True(t, bounds[0].MinValidPrice.EQ(t1lb1))
+	require.True(t, bounds[0].MaxValidPrice.EQ(t1ub1))
+	require.Equal(t, bounds[0].ReferencePrice, decPrice)
+	require.Equal(t, *bounds[1].Trigger.IntoProto(), t2)
+	require.True(t, bounds[1].MinValidPrice.EQ(t2lb1))
+	require.True(t, bounds[1].MaxValidPrice.EQ(t2ub1))
+	require.Equal(t, bounds[1].ReferencePrice, decPrice)
+
+	end := types.AuctionDuration{Duration: t1.AuctionExtension}
+	pm.SetMinDuration(time.Duration(end.Duration) * time.Second)
+	auctionStateMock.EXPECT().StartPriceAuction(now, &end).Times(1)
+
+	cPrice := num.Sum(price1, maxUp2)
+	cPrice.Sub(cPrice, maxUp1)
+	cp2 := []*types.Trade{{Price: cPrice, Size: 1}}
+	b = pm.CheckPrice(ctx, auctionStateMock, cp2, true) // t1 violated only
+	require.False(t, b)
+
+	initialAuctionEnd := now.Add(time.Duration(t1.AuctionExtension) * time.Second)
+
+	auctionStateMock.EXPECT().IsFBA().Return(false).Times(1)
+	auctionStateMock.EXPECT().InAuction().Return(true).Times(1)
+	auctionStateMock.EXPECT().IsOpeningAuction().Return(false).Times(1)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(true).AnyTimes()
+	auctionStateMock.EXPECT().ExpiresAt().Return(&initialAuctionEnd).Times(1)
+
+	bounds = pm.GetCurrentBounds()
+	require.Len(t, bounds, 1)
+	require.Equal(t, *bounds[0].Trigger.IntoProto(), t2)
+	require.True(t, bounds[0].MinValidPrice.EQ(t2lb1))
+	require.True(t, bounds[0].MaxValidPrice.EQ(t2ub1))
+	require.Equal(t, bounds[0].ReferencePrice, decPrice)
+
+	afterInitialAuction := initialAuctionEnd.Add(time.Nanosecond)
+	now = afterInitialAuction
+
+	auctionStateMock.EXPECT().SetReadyToLeave().Times(1)
+
+	cPrice = num.Sum(price1, maxUp2, maxUp1)
+	pm.OnTimeUpdate(afterInitialAuction)
+	cp3 := []*types.Trade{{Price: cPrice, Size: 1}}
+	b = pm.CheckPrice(ctx, auctionStateMock, cp3, true) // price should violated 2nd trigger and result in auction extension
+	require.False(t, b)
+}
+
 func TestMarketInOpeningAuction(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
@@ -804,7 +916,7 @@ func TestMarketInOpeningAuction(t *testing.T) {
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -839,12 +951,13 @@ func TestMarketInGenericAuction(t *testing.T) {
 	auctionStateMock.EXPECT().InAuction().Return(true).Times(5)
 	auctionStateMock.EXPECT().IsOpeningAuction().Return(false).Times(5)
 	auctionStateMock.EXPECT().IsPriceAuction().Return(false).AnyTimes()
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).AnyTimes()
 	auctionStateMock.EXPECT().IsPriceExtension().Return(false).AnyTimes()
 	auctionStateMock.EXPECT().CanLeave().Return(false).AnyTimes()
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -898,7 +1011,7 @@ func TestGetValidPriceRange_NoTriggers(t *testing.T) {
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 
@@ -942,10 +1055,12 @@ func TestGetValidPriceRange_2triggers(t *testing.T) {
 	currentPriceD := currentPrice.ToDecimal()
 	auctionStateMock.EXPECT().IsFBA().Return(false).Times(12)
 	auctionStateMock.EXPECT().InAuction().Return(false).Times(12)
+	auctionStateMock.EXPECT().IsPriceAuction().Return(false).Times(11)
+	auctionStateMock.EXPECT().IsLiquidityAuction().Return(false).Times(11)
 	statevar := mocks.NewMockStateVarEngine(ctrl)
 	statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any())
 
-	pm, err := price.NewMonitor("asset", "market", riskModel, settings, statevar, logging.NewTestLogger())
+	pm, err := price.NewMonitor("asset", "market", riskModel, auctionStateMock, settings, statevar, logging.NewTestLogger())
 	require.NoError(t, err)
 	require.NotNil(t, pm)
 	downFactors := []num.Decimal{pMin1.Div(currentPriceD), pMin2.Div(currentPriceD)}